Remove x64 build warning.

+++ plugin/Plugin/ScriptablePluginObjectBase.cpp
@@ -197,7 +197,7 @@ CString NPStringToCString(NPString npstr)
 CString NPIdentifierToCString(NPIdentifier npid)
 {
 	NPUTF8* putf8IdName = NPN_UTF8FromIdentifier(npid);
-	NPString npstrIdName = { putf8IdName, strlen(putf8IdName) };
+	NPString npstrIdName = { putf8IdName, (uint32_t)strlen(putf8IdName) };
 	CString idName = NPStringToCString(npstrIdName);
 	NPN_MemFree(putf8IdName);
 	return idName;
